Use requestRTT function on all measure tests
